HTML/CSS stuff
This commit looks larger than it is.

This is just some simple html/css I worked on today. I haven't been
using bootstrap yet, and most of these are just blank pages with a
couple icon links. I created a readme file pertaining to the file
directorial setup as of right now.

So far, everything is non-functional and is just serving as a template
to work off of when incorporating some functionality. None of what I am
committing is in any way finished.