Countdown timer w/ hardcoded start time
The timer is using a hardcoded start time at the moment. Changes to
myCtrl are just commented out code that I was working with to try to get
it to work there. Changes to app.js is a new time controller that is
also taking care of the food route due to the timer needing access to
that data. I couldn't get the timer to work in myCtrl.js for whatever
reason, but it works as it is now. I commented out geolocate and
geocoder requirements from sandbox because I couldn't start the server
with those there.